Added Python 3 compatibility for unquote()
unquote() can now be used with either Python2 or Python 3.  unquote() would cause the Python3 firmware on my esp32 to crash on a NameError, because it was calling xrange() instead of range().  Two functions were created to handle either Python 2 or Python 3.  If a NameError is thrown, _url_decode_py3() will be called.